requirements.txt
python-telegram-bot
python-telegram-bot-pagination
requests
pytz